Start xBehaviour; Remove zScript.s
I need sleep.
